+++ Providers/EFCore/Revo.EFCore/Repositories/EFCoreRepositoriesModule.cs
@@ -1,15 +1,6 @@
-﻿using System;
-using Ninject.Extensions.ContextPreservation;
-using Ninject.Modules;
-using Revo.Core.Commands;
+﻿using Ninject.Modules;
 using Revo.Core.Core;
-using Revo.Core.Transactions;
-using Revo.DataAccess.Entities;
 using Revo.EFCore.Configuration;
-using Revo.EFCore.DataAccess.Conventions;
-using Revo.EFCore.DataAccess.Entities;
-using Revo.EFCore.Domain;
-using Revo.EFCore.Projections;
 using Revo.EFCore.Repositories;
 using Revo.Infrastructure.Repositories;