+++ Revo.Infrastructure/EventStores/Generic/EventStore.cs
@@ -4,7 +4,6 @@ using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using MoreLinq;
 using Revo.Core.Core;
 using Revo.Core.Types;
 using Revo.DataAccess.Entities;